also check for mongo.exe

+++ .evergreen/download-mongodb.sh
@@ -515,7 +515,7 @@ download_and_extract ()
 
    download_and_extract_package "$MONGODB_DOWNLOAD_URL" "$EXTRACT"
 
-   if [ ! -e $DRIVERS_TOOLS/mongodb/bin/mongo ]; then
+   if [ ! -e $DRIVERS_TOOLS/mongodb/bin/mongo -a ! -e $DRIVERS_TOOLS/mongodb/bin/mongo.exe ]; then
       # The legacy mongo shell is not included in server downloads of 6.0.0-rc6 or later. Refer: SERVER-64352.
       # Some test scripts use the mongo shell for setup.
       # Download 5.0 package to get the legacy mongo shell as a workaround until DRIVERS-2328 is addressed.
@@ -527,7 +527,11 @@ download_and_extract ()
       mkdir $DRIVERS_TOOLS/legacy-shell-download
       DRIVERS_TOOLS=$DRIVERS_TOOLS/legacy-shell-download
       download_and_extract_package "$MONGODB_DOWNLOAD_URL" "$EXTRACT"
-      cp $DRIVERS_TOOLS/mongodb/bin/mongo $SAVED_DRIVERS_TOOLS/mongodb/bin
+      if [ -e $DRIVERS_TOOLS/mongodb/bin/mongo ]; then
+         cp $DRIVERS_TOOLS/mongodb/bin/mongo $SAVED_DRIVERS_TOOLS/mongodb/bin
+      elif [ -e $DRIVERS_TOOLS/mongodb/bin/mongo.exe ]; then
+         cp $DRIVERS_TOOLS/mongodb/bin/mongo.exe $SAVED_DRIVERS_TOOLS/mongodb/bin
+      fi
       DRIVERS_TOOLS=$SAVED_DRIVERS_TOOLS
       rm -rf $DRIVERS_TOOLS/legacy-shell-download
       echo "Download legacy shell from 5.0 ... end"
